refactor: share prompt runner parsing
